src: introduce link handle
Abstract the file descriptor handle to a more abstract type so it's
possible to use different 'backends' to send/receive commands, e.g.
via MI instead direct.

The name 'link' was chosen because this name is not in use and thus we
avoid overloading existing names, such ctrl. The specification is using
'link' to describe the connection on a PCI bus which is kind of matching
how the handle is used in the library. It is the means to talk to a
controller.

Signed-off-by: Daniel Wagner <wagi@kernel.org>